requirements.txt
# requirements.txt

addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
blessed==1.20.0
bounding-box==0.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
easydict==1.13
efficientnet-pytorch==0.7.1
fairscale==0.4.13
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
gpustat==1.1.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.4.0
lightning-utilities==0.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mecab-python3==1.0.9
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.6
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
propcache==0.2.1
psutil==6.1.1
pycocotools==2.0.8
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.1.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scipy==1.15.1
segmentation-models-pytorch==0.3.3
shapely==2.0.7
six==1.17.0
sympy==1.13.1
termcolor==2.5.0
terminaltables==3.1.10
timm==0.9.2
tokenizers==0.15.2
tomli==2.2.1
torch==2.1.0+cu121
# torch==2.1.0
# torch==2.8.0+cu128 설치함
# python -c "import torch; print(torch.__version__)"
torchmetrics==1.6.1
torchvision==0.16.0+cu121
# torchvision==0.16.0
# torchvision==0.23.0+cu128 설치함
# python -c "import torchvision; print(torchvision.__version__)"
tqdm==4.67.1
transformers==4.37.0
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yapf==0.43.0
yarl==1.18.3
